pass content from components/page
