+++ 05_フロントエンド/landing-page/src/components/template/Home/Home.tsx
@@ -19,80 +19,21 @@ import {
 } from "@/components/organism/CardPostSmall"
 import { Pagenation } from "@/components/organism/Pagenation"
 
-const _img = (image?: string): string | undefined =>
-  image ? `/image/${image}` : undefined
-
-const _genPost = (title: string, image?: string): CardPostProps => ({
-  date: new Date(),
-  tag: {
-    name: "Larabel",
-    link: "",
-  },
-  title,
-  body: "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora expedita dicta totam aspernatur doloremque. Excepturi iste iusto eos enim reprehenderit nisi, accusamus delectus nihil quis facere in modi ratione libero!",
-  link: "/#",
-  author: {
-    name: "Alex John",
-    image: _img(image),
-  },
-})
-
-const _genAuthor = (name: string, image?: string): AuthorDetailsProps => ({
-  name,
-  image: _img(image),
-  numberOfPosts: 20,
-})
-
-const _genCategory = (name: string): _Category => ({
-  name,
-  link: "/#",
-})
-
 type _Category = {
   name: string
   link: string
 }
 
-export const Home = (): JSX.Element => {
-  const posts: CardPostProps[] = [
-    _genPost("Build Your New Idea with Laravel Freamwork."),
-    _genPost("Accessibility tools for designers and developers"),
-    _genPost("PHP: Array to Map"),
-    _genPost("Django Dashboard - Learn by Coding"),
-    _genPost("TDD First"),
-  ]
+export type HomeTemplateProps = {
+  posts: CardPostProps[]
+  authors: AuthorDetailsProps[]
+  categories: _Category[]
+  resentPost: CardPostSmallProps
+  options: OptionItem[]
+}
 
-  const authors: AuthorDetailsProps[] = [
-    _genAuthor("Alex John"),
-    _genAuthor("Jane Doe", "neko.png"),
-    _genAuthor("Lisa Way"),
-    _genAuthor("Steve Matt", "sushidesu.jpg"),
-    _genAuthor("Khatab Wedaa"),
-  ]
-
-  const categories: _Category[] = [
-    _genCategory("AWS"),
-    _genCategory("Laravel"),
-    _genCategory("Vue"),
-    _genCategory("Design"),
-    _genCategory("Django"),
-    _genCategory("PHP"),
-  ]
-
-  const resentPost: CardPostSmallProps = _genPost(
-    "Build Your New Idea with Laravel Freamwork."
-  )
-
-  const options: OptionItem[] = [
-    {
-      label: "Latest",
-      value: "latest",
-    },
-    {
-      label: "Last Week",
-      value: "last-week",
-    },
-  ]
+export const Home = (props: HomeTemplateProps): JSX.Element => {
+  const { posts, authors, categories, resentPost, options } = props
 
   return (
     <div className={styles["wrapper"]}>
