+++ 05_フロントエンド/landing-page/src/components/template/Home/index.ts
@@ -1,3 +1,4 @@
 import { HomePage } from "./Home.page"
+export * from "./Home"
 
 export default HomePage
